Implements the WpsServer interface into a dedicated class WpsServerImplementation wich will be extended by the LocalWpsServerImpl class.
